+++ src/App.jsx
@@ -4,6 +4,7 @@ import { Header } from './components/Header/header'
 import { Home } from "./pages/Home"
 import { Login } from './pages/login'
 import { BlueBotton } from './pages/BlueBotton'
+import { Footer } from './components/footer/footer'
 
 
 function App() {
@@ -11,11 +12,14 @@ function App() {
   return (
     <Router>
       <Header/>
+
       <Routes>
         <Route path="/" element={<Home/>} />
         <Route path="/login" element={<Login/>} />
         <Route path="/bluebutton" element={<BlueBotton/>} />
       </Routes>
+
+      <Footer/>
     </Router>
   )
 }
